+++ src/compas_timber/connections/l_french_ridge_lap.py
@@ -1,16 +1,13 @@
-from compas.tolerance import TOL
-
-from compas_timber.connections.utilities import beam_ref_side_incidence
-from compas_timber.connections.utilities import beam_ref_side_incidence_with_vector
 from compas_timber.errors import BeamJoiningError
 from compas_timber.fabrication import FrenchRidgeLap
 
-from .joint import Joint
+from .lap_joint import LapJoint
 from .solver import JointTopology
+from .utilities import are_beams_coplanar
 
 
-class LFrenchRidgeLapJoint(Joint):
-    """Represents an L-FrenchRidgeLap type joint which joins two beams in their ends, by lapping them with a ridge.
+class LFrenchRidgeLapJoint(LapJoint):
+    """Represents an L-FrenchRidgeLap type joint which joins two beams at their ends, by lapping them with a ridge.
     The joint can only be created between two beams that are aligned and have the same dimensions.
 
     This joint type is compatible with beams in L topology.
@@ -19,89 +16,39 @@ class LFrenchRidgeLapJoint(Joint):
 
     Parameters
     ----------
-    beam_a : :class:`~compas_timber.parts.Beam`
-        First beam to be joined.
-    beam_b : :class:`~compas_timber.parts.Beam`
-        Second beam to be joined.
+    main_beam : :class:`~compas_timber.elements.Beam`
+        The main beam to be joined.
+    cross_beam : :class:`~compas_timber.elements.Beam`
+        The cross beam to be joined.
+    flip_lap_side : bool
+        If True, the lap is flipped to the other side of the beams.
     drillhole_diam : float
         Diameter of the drill hole to be made in the joint.
-    flip_beams : bool
-        If True, the beams will be flipped in the joint. Default is False.
 
     Attributes
     ----------
-    beam_a : :class:`~compas_timber.parts.Beam`
-        First beam to be joined.
-    beam_b : :class:`~compas_timber.parts.Beam`
-        Second beam to be joined.
+    main_beam : :class:`~compas_timber.elements.Beam`
+        The main beam to be joined.
+    cross_beam : :class:`~compas_timber.elements.Beam`
+        The cross beam to be joined.
+    flip_lap_side : bool
+        If True, the lap is flipped to the other side of the beams.
     drillhole_diam : float
         Diameter of the drill hole to be made in the joint.
-    flip_beams : bool
-        If True, the beams will be flipped in the joint. Default is False.
 
     """
 
     SUPPORTED_TOPOLOGY = JointTopology.TOPO_L
 
-    @property
-    def __data__(self):
-        data = super(LFrenchRidgeLapJoint, self).__data__
-        data["beam_a_guid"] = self.beam_a_guid
-        data["beam_b_guid"] = self.beam_b_guid
-        data["drillhole_diam"] = self.drillhole_diam
-        data["flip_beams"] = self.flip_beams
-        return data
-
-    def __init__(self, beam_a=None, beam_b=None, drillhole_diam=None, flip_beams=None, **kwargs):
-        super(LFrenchRidgeLapJoint, self).__init__(**kwargs)
-        self.beam_a = beam_a
-        self.beam_b = beam_b
-        self.beam_a_guid = kwargs.get("beam_a_guid", None) or str(beam_a.guid)
-        self.beam_b_guid = kwargs.get("beam_b_guid", None) or str(beam_b.guid)
+    def __init__(self, main_beam=None, cross_beam=None, flip_lap_side=False, drillhole_diam=None, **kwargs):
+        super(LFrenchRidgeLapJoint, self).__init__(main_beam, cross_beam, flip_lap_side, drillhole_diam, **kwargs)
 
         self.drillhole_diam = drillhole_diam
-        self.flip_beams = flip_beams
-        self.features = []
-
-    @property
-    def elements(self):
-        return [self.beam_a, self.beam_b]
-
-    @property
-    def beam_a_ref_side_index(self):
-        cross_vector = self.beam_a.centerline.direction.cross(self.beam_b.centerline.direction)
-        ref_side_dict = beam_ref_side_incidence_with_vector(self.beam_a, cross_vector, ignore_ends=True)
-        if self.flip_beams:
-            return max(ref_side_dict, key=ref_side_dict.get)
-        return min(ref_side_dict, key=ref_side_dict.get)
-
-    @property
-    def beam_b_ref_side_index(self):
-        cross_vector = self.beam_a.centerline.direction.cross(self.beam_b.centerline.direction)
-        ref_side_dict = beam_ref_side_incidence_with_vector(self.beam_b, cross_vector, ignore_ends=True)
-        if self.flip_beams:
-            return min(ref_side_dict, key=ref_side_dict.get)
-        return max(ref_side_dict, key=ref_side_dict.get)
-
-    @property
-    def cutting_plane_a(self):
-        # the plane that cuts beam_b
-        ref_side_dict = beam_ref_side_incidence(self.beam_b, self.beam_a, ignore_ends=True)
-        ref_side_index = max(ref_side_dict, key=ref_side_dict.get)
-        return self.beam_a.ref_sides[ref_side_index]
-
-    @property
-    def cutting_plane_b(self):
-        # the plane that cuts beam_a
-        ref_side_dict = beam_ref_side_incidence(self.beam_a, self.beam_b, ignore_ends=True)
-        ref_side_index = max(ref_side_dict, key=ref_side_dict.get)
-        return self.beam_b.ref_sides[ref_side_index]
 
     def add_extensions(self):
         """Calculates and adds the necessary extensions to the beams.
 
-        This method is called during the `Model.process_joinery()` process after the joint
-        has been instantiated and added to the model.
+        This method is automatically called when joint is created by the call to `Joint.create()`.
 
         Raises
         ------
@@ -109,20 +56,20 @@ class LFrenchRidgeLapJoint(Joint):
             If the extension could not be calculated.
 
         """
+        assert self.main_beam and self.cross_beam
 
-        assert self.beam_a and self.beam_b
-        start_a, start_b = None, None
+        start_main, start_cross = None, None
         try:
-            start_a, end_a = self.beam_a.extension_to_plane(self.cutting_plane_b)
-            start_b, end_b = self.beam_b.extension_to_plane(self.cutting_plane_a)
+            start_main, end_main = self.main_beam.extension_to_plane(self.main_cutting_plane)
+            start_cross, end_cross = self.cross_beam.extension_to_plane(self.cross_cutting_plane)
         except AttributeError as ae:
             # I want here just the plane that caused the error
-            geometries = [self.cutting_plane_a] if start_a is not None else [self.cutting_plane_b]
+            geometries = [self.cross_cutting_plane] if start_main is not None else [self.main_cutting_plane]
             raise BeamJoiningError(self.elements, self, debug_info=str(ae), debug_geometries=geometries)
         except Exception as ex:
             raise BeamJoiningError(self.elements, self, debug_info=str(ex))
-        self.beam_a.add_blank_extension(start_a, end_a, self.guid)
-        self.beam_b.add_blank_extension(start_b, end_b, self.guid)
+        self.main_beam.add_blank_extension(start_main, end_main, self.main_beam_guid)
+        self.cross_beam.add_blank_extension(start_cross, end_cross, self.cross_beam_guid)
 
     def add_features(self):
         """Adds the necessary features to the beams.
@@ -132,38 +79,39 @@ class LFrenchRidgeLapJoint(Joint):
         have been added via `Joint.add_extensions()`.
 
         """
-        assert self.beam_a and self.beam_b
+        assert self.main_beam and self.cross_beam
 
         if self.features:
-            self.beam_a.remove_features(self.features)
-            self.beam_b.remove_features(self.features)
+            self.main_beam.remove_features(self.features)
+            self.cross_beam.remove_features(self.features)
 
-        frl_a = FrenchRidgeLap.from_beam_beam_and_plane(self.beam_a, self.beam_b, self.cutting_plane_b, self.drillhole_diam, self.beam_a_ref_side_index)
-        frl_b = FrenchRidgeLap.from_beam_beam_and_plane(self.beam_b, self.beam_a, self.cutting_plane_a, self.drillhole_diam, self.beam_b_ref_side_index)
-        self.beam_a.add_features(frl_a)
-        self.beam_b.add_features(frl_b)
-        self.features = [frl_a, frl_b]
+        main_frl_feature = FrenchRidgeLap.from_beam_beam_and_plane(self.main_beam, self.cross_beam, self.main_cutting_plane, self.drillhole_diam, self.main_ref_side_index)
+        cross_frl_feature = FrenchRidgeLap.from_beam_beam_and_plane(self.cross_beam, self.main_beam, self.cross_cutting_plane, self.drillhole_diam, self.cross_ref_side_index)
+        # store the features to the beams
+        self.main_beam.add_features(main_frl_feature)
+        self.cross_beam.add_features(cross_frl_feature)
+        # register the features in the joint
+        self.features = [main_frl_feature, cross_frl_feature]
 
     def check_elements_compatibility(self):
         """Checks if the elements are compatible for the creation of the joint.
+
+        Compared to the LapJoint's `check_elements_compatibility` method, this one additionally checks if dimensions of the beams match.
 
         Raises
         ------
         BeamJoiningError
             If the elements are not compatible for the creation of the joint.
-
         """
-        # check if the beams are aligned
-        cross_vect = self.beam_a.centerline.direction.cross(self.beam_b.centerline.direction)
-        for beam in self.elements:
-            beam_normal = beam.frame.normal.unitized()
-            dot = abs(beam_normal.dot(cross_vect.unitized()))
-            if not (TOL.is_zero(dot) or TOL.is_close(dot, 1)):
-                raise BeamJoiningError(self.elements, self, debug_info="The two beams are not aligned to create a French Ridge Lap joint.")
-
+        if not are_beams_coplanar(*self.elements):
+            raise BeamJoiningError(
+                beams=self.elements,
+                joint=self,
+                debug_info="The two beams are not coplanar to create a Lap joint.",
+            )
         # calculate widths and heights of the beams
         dimensions = []
-        ref_side_indices = [self.beam_a_ref_side_index, self.beam_b_ref_side_index]
+        ref_side_indices = [self.main_ref_side_index, self.cross_ref_side_index]
         for i, beam in enumerate(self.elements):
             width = beam.side_as_surface(ref_side_indices[i]).ysize
             height = beam.height if ref_side_indices[i] % 2 == 0 else beam.width