+++ src/compas_timber/fabrication/btlx.py
@@ -607,16 +607,29 @@ class MachiningLimits(object):
         Limit the front face.
     face_limited_back : bool
         Limit the back face.
+    face_limited_top : bool
+        Limit the top face.
+    face_limited_bottom : bool
+        Limit the bottom face.
 
     """
 
-    EXPECTED_KEYS = ["FaceLimitedStart", "FaceLimitedEnd", "FaceLimitedFront", "FaceLimitedBack"]
+    EXPECTED_KEYS = [
+        "FaceLimitedStart",
+        "FaceLimitedEnd",
+        "FaceLimitedFront",
+        "FaceLimitedBack",
+        "FaceLimitedTop",
+        "FaceLimitedBottom",
+    ]
 
     def __init__(self):
         self.face_limited_start = True
         self.face_limited_end = True
         self.face_limited_front = True
         self.face_limited_back = True
+        self.face_limited_top = True
+        self.face_limited_bottom = True
 
     @property
     def limits(self):
@@ -626,6 +639,8 @@ class MachiningLimits(object):
             "FaceLimitedEnd": self.face_limited_end,
             "FaceLimitedFront": self.face_limited_front,
             "FaceLimitedBack": self.face_limited_back,
+            "FaceLimitedTop": self.face_limited_top,
+            "FaceLimitedBottom": self.face_limited_bottom,
         }