+++ src/compas_timber/ghpython/components/CT_Joint_Rule_Topology/code.py
@@ -3,7 +3,7 @@ from ghpythonlib.componentbase import executingcomponent as component
 from compas_timber.connections import JointTopology
 from compas_timber.connections import LMiterJoint
 from compas_timber.connections import TButtJoint
-from compas_timber.connections import XHalfLapJoint
+from compas_timber.connections import XLapJoint
 from compas_timber.design import TopologyRule
 
 
@@ -12,6 +12,6 @@ class DefaultJointRule(component):
         topoRules = []
         topoRules.append(TopologyRule(JointTopology.TOPO_L, LMiterJoint))
         topoRules.append(TopologyRule(JointTopology.TOPO_T, TButtJoint))
-        topoRules.append(TopologyRule(JointTopology.TOPO_X, XHalfLapJoint))
+        topoRules.append(TopologyRule(JointTopology.TOPO_X, XLapJoint))
 
         return topoRules